Three functions separated.
Password toggle & Confirm Password toggle & Email Validation functions separated.
